+++ model/src/main/java/up/TowerDefense/model/object/Wall.java
@@ -1,15 +1,13 @@
 package up.TowerDefense.model.object;
 
-import java.awt.image.BufferedImage;
-
 public class Wall extends PlaceableObstacle{
 
     private static int newHealthScale = 5;
-    private static double newBuyingCost = 100;
+    private static int newBuyingCost = 100;
     private static float newSize = 1.0f;
     private static String image;
 
-    public Wall(double x, double y) {
+    public Wall(float x, float y) {
         super(x, y, newSize, newHealthScale, newHealthScale, ObsType.WALL, newBuyingCost,image);
 
     }
